+++ apps/api/src/routes/legacy_keys_createKey.test.ts
@@ -27,7 +27,7 @@ describe("simple", () => {
       },
     });
 
-    expect(res.status).toEqual(200);
+    expect(res.status, `expected 200, received: ${JSON.stringify(res)}`).toBe(200);
 
     const found = await h.db.readonly.query.keys.findFirst({
       where: (table, { eq }) => eq(table.id, res.body.keyId),
@@ -81,7 +81,7 @@ describe("with prefix", () => {
       },
     });
 
-    expect(res.status).toEqual(200);
+    expect(res.status, `expected 200, received: ${JSON.stringify(res)}`).toBe(200);
 
     const key = await h.db.readonly.query.keys.findFirst({
       where: (table, { eq }) => eq(table.id, res.body.keyId),