+++ apps/api/src/routes/v1_keys_createKey.security.test.ts
@@ -76,7 +76,7 @@ describe("correct roles", () => {
           apiId,
         },
       });
-      expect(res.status).toEqual(200);
+      expect(res.status, `expected 200, received: ${JSON.stringify(res)}`).toBe(200);
     });
   });
 });