+++ packages/api/src/openapi.d.ts
@@ -40,6 +40,9 @@ export interface paths {
   "/v1/apis.deleteApi": {
     post: operations["deleteApi"];
   };
+  "/v1/apis.deleteKeys": {
+    post: operations["deleteKeys"];
+  };
   "/v1/ratelimits.limit": {
     post: operations["limit"];
   };
@@ -264,7 +267,7 @@ export interface components {
        * @description The unix timestamp in milliseconds when the key was created
        * @example 0
        */
-      createdAt?: number;
+      createdAt: number;
       /**
        * @description The unix timestamp in milliseconds when the key was last updated
        * @example 0
@@ -354,6 +357,8 @@ export interface components {
        * @example true
        */
       enabled?: boolean;
+      /** @description The key in plaintext */
+      plaintext?: string;
     };
     V1KeysVerifyKeyResponse: {
       /**
@@ -592,6 +597,7 @@ export interface operations {
     parameters: {
       query: {
         keyId: string;
+        decrypt?: boolean | null;
       };
     };
     responses: {
@@ -1398,6 +1404,7 @@ export interface operations {
         limit?: number;
         cursor?: string;
         ownerId?: string;
+        decrypt?: boolean | null;
       };
     };
     responses: {
@@ -1477,6 +1484,77 @@ export interface operations {
       200: {
         content: {
           "application/json": Record<string, never>;
+        };
+      };
+      /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
+      400: {
+        content: {
+          "application/json": components["schemas"]["ErrBadRequest"];
+        };
+      };
+      /** @description Although the HTTP standard specifies "unauthorized", semantically this response means "unauthenticated". That is, the client must authenticate itself to get the requested response. */
+      401: {
+        content: {
+          "application/json": components["schemas"]["ErrUnauthorized"];
+        };
+      };
+      /** @description The client does not have access rights to the content; that is, it is unauthorized, so the server is refusing to give the requested resource. Unlike 401 Unauthorized, the client's identity is known to the server. */
+      403: {
+        content: {
+          "application/json": components["schemas"]["ErrForbidden"];
+        };
+      };
+      /** @description The server cannot find the requested resource. In the browser, this means the URL is not recognized. In an API, this can also mean that the endpoint is valid but the resource itself does not exist. Servers may also send this response instead of 403 Forbidden to hide the existence of a resource from an unauthorized client. This response code is probably the most well known due to its frequent occurrence on the web. */
+      404: {
+        content: {
+          "application/json": components["schemas"]["ErrNotFound"];
+        };
+      };
+      /** @description This response is sent when a request conflicts with the current state of the server. */
+      409: {
+        content: {
+          "application/json": components["schemas"]["ErrConflict"];
+        };
+      };
+      /** @description The user has sent too many requests in a given amount of time ("rate limiting") */
+      429: {
+        content: {
+          "application/json": components["schemas"]["ErrTooManyRequests"];
+        };
+      };
+      /** @description The server has encountered a situation it does not know how to handle. */
+      500: {
+        content: {
+          "application/json": components["schemas"]["ErrInternalServerError"];
+        };
+      };
+    };
+  };
+  deleteKeys: {
+    requestBody: {
+      content: {
+        "application/json": {
+          /**
+           * @description The id of the api, that the keys belong to.
+           * @example api_1234
+           */
+          apiId: string;
+          /**
+           * @description If true, the keys will be permanently deleted. If false, the keys will be soft-deleted and can be restored later.
+           * @default false
+           */
+          permanent?: boolean;
+        };
+      };
+    };
+    responses: {
+      /** @description The keys have been deleted */
+      200: {
+        content: {
+          "application/json": {
+            /** @description The number of keys that were deleted */
+            deletedKeys: number;
+          };
         };
       };
       /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
@@ -1690,7 +1768,10 @@ export interface operations {
            * @example my key
            */
           name?: string;
-          hash: {
+          /** @description The raw key in plaintext. If provided, unkey encrypts this value and stores it securely. Provide either `hash` or `plaintext` */
+          plaintext?: string;
+          /** @description Provide either `hash` or `plaintext` */
+          hash?: {
             /** @description The hashed and encoded key */
             value: string;
             /**
@@ -1701,7 +1782,6 @@ export interface operations {
           };
           /**
            * @description The first 4 characters of the key. If a prefix is used, it should be the prefix plus 4 characters.
-           * @default
            * @example unkey_32kq
            */
           start?: string;