+++ packages/api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @unkey/api
 
+## 0.20.4
+
+### Patch Changes
+
+- 3ddb1f1: feat: apis.deleteKeys
+
+## 0.20.3
+
+### Patch Changes
+
+- fc2b651: update
+- Updated dependencies [fc2b651]
+  - @unkey/rbac@0.1.13
+
 ## 0.20.2
 
 ### Patch Changes